1、优化标签、分类的删除功能 2、前台:标签随机背景、iconfont字体等

+++ src/main/java/com/lzt/ssm/blog/controller/admin/BackCategoryController.java
@@ -58,7 +58,11 @@ public class BackCategoryController {
 
     @RequestMapping("/delete/{id}")
     public String delete(@PathVariable("id") Integer id) {
-        categoryService.deleteEntityById(id);
+        Integer count = categoryService.countArticleByCategoryId(id);
+        if(count == 0){
+            categoryService.deleteEntityById(id);
+        }
+
         return "redirect:/admin/category";
     }
